+++ marshmallow/compat.py
@@ -37,27 +37,6 @@ else:
     from collections import OrderedDict
     OrderedDict = OrderedDict
 
-def plain_function(f):
-    """Ensure that ``callable`` is a plain function rather than an unbound method, as
-    may happen in Python 2 when setting functions as class variables. This is particularly
-    useful for class Meta options that are functions. ::
-
-        class MySchemaOpts(SchemaOpts):
-            def __init__(self, meta):
-                self.some_option = plain_function(getattr(meta, 'some_option', None))
-
-        class MySchema(Schema):
-            OPTIONS_CLASS = MySchemaOpts
-
-            class Meta:
-                some_option = my_function
-
-    Returns `None` if callable is `None`.
-    """
-    if PY2 and f:
-        return f.im_func
-    # Python 3 doesn't have bound/unbound methods, so don't need to do anything
-    return f
 
 # From six
 def with_metaclass(meta, *bases):